refactor: Consolidate theme handling and update community version announcement logic
