+++ PCL.Core/App/States.cs
@@ -236,21 +236,6 @@ public static partial class States
         /// </summary>
         [ConfigItem<double>("WindowWidth", 900, ConfigSource.Local)] public partial double WindowWidth { get; set; }
 
-        /// <summary>
-        /// 传说中的秋仪金代码，但是没卵用。
-        /// </summary>
-        [ConfigItem<string>("UiLauncherThemeGold", "")] public partial string ThemeGoldCode { get; set; }
-
-        /// <summary>
-        /// 传说中的隐藏主题1，但是没卵用
-        /// </summary>
-        [ConfigItem<string>("UiLauncherThemeHide", "0|1|2|3|4")] public partial string ThemeHiddenV1 { get; set; }
-
-        /// <summary>
-        /// 传说中的隐藏主题2，但是没卵用。
-        /// </summary>
-        [ConfigItem<string>("UiLauncherThemeHide2", "0|1|2|3|4")] public partial string ThemeHiddenV2 { get; set; }
-
     }
 
     /// <summary>
